- Get rid of server file - Bring code from server back into index and eliminate the function - Export the server const
